requirements.txt
flask
matplotlib
flask
biopython
jinja2
